Refactoring sliceutil
